requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
Deprecated==1.2.14
fastapi==0.103.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
mdx-linkify==2.1
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyGithub==1.59.1
PyJWT==2.8.0
PyNaCl==1.5.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
webencodings==0.5.1
wrapt==1.15.0
